Create a fallback R environment for shared library problems. (#6)
This aims to solve GLIBCXX versioning errors, where certain Python packages
(i.e., scipy) are compiled against one version of libstdc++ but R is compiled
against a different version.  When the two are forced to link to each other by
reticulate, the Python packages fail to import with a GLIBCXX version error. 

We solve this by creating a conda environment with a minimal R + reticulate
installation, hoping that that conda is self-consistent with respect to the
libstdc++ version it uses across Python and R builds. basilisk can then perform
its reticulate calls inside this "fallback" environment if all else fails.

For most users, this environment is only created when GLIBCXX errors are
detected, thus avoiding the overhead of another R installation unless it is
truly needed. System installs don't have a choice and always create a fallback
environment unless explicitly instructed not to by 'noFallbackR()'.